@@ -1,12 +1,12 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdk 32
+    compileSdk 33
     buildToolsVersion '33.0.0'
 
     defaultConfig {
         minSdk 24
-        targetSdk 32
+        targetSdk 33
         versionCode 4
         versionName "1.0.3"
         consumerProguardFiles "proguard-consumer-rules.pro"
