@@ -3,12 +3,12 @@ plugins {
 }
 
 android {
-    compileSdk 32
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.matthew.unscramble"
         minSdk 24
-        targetSdk 32
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -28,14 +28,8 @@ android {
 }
 
 dependencies {
-    implementation 'androidx.room:room-runtime:2.5.0-alpha02'
+    implementation 'com.google.code.gson:gson:2.9.1'
     implementation project(path: ':animcubeandroid')
-    annotationProcessor "androidx.room:room-compiler:2.4.3"
-    implementation "androidx.room:room-rxjava2:2.4.3"
-    implementation "androidx.room:room-rxjava3:2.4.3"
-    implementation "androidx.room:room-guava:2.4.3"
-    testImplementation "androidx.room:room-testing:2.4.3"
-    implementation "androidx.room:room-paging:2.5.0-alpha02"
     implementation 'androidx.appcompat:appcompat:1.5.0'
     implementation 'com.google.android.material:material:1.6.1'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
