@@ -181,7 +181,7 @@ public class AnimCube extends SurfaceView implements View.OnTouchListener {
     private boolean showBackFaces;
     private double faceShift;
     // move sequence data
-    private int[] move;
+    public int[] move;
     private int movePos;
     private int moveDir;
     private boolean moveOne;
@@ -1184,7 +1184,8 @@ public class AnimCube extends SurfaceView implements View.OnTouchListener {
                 }
             } else {
                 for (int j = 0; j < 21; j++) {
-                    if (sequence.charAt(i) == "UDFBLRESMXYZxyzudfblr".charAt(j)) {
+                    //Modified so that Z and Y moves are correct
+                    if (sequence.charAt(i) == "UDFBLRESMXZYxzyudfblr".charAt(j)) {
                         i++;
                         int mode = moveModes[j];
                         move[length] = moveCodes[j] * 24;
@@ -1332,9 +1333,12 @@ public class AnimCube extends SurfaceView implements View.OnTouchListener {
                 animating = true;
                 int[] mv = move;
                 if (moveDir > 0) {
+                    /*
                     if (movePos >= mv.length) {
                         movePos = 0;
                     }
+                    */
+                    movePos = 0;
                 } else {
                     if (movePos == 0) {
                         movePos = mv.length;
